updated type export
