+++ packages/rag/src/document/schema/index.ts
@@ -1,3 +1,3 @@
 export { BaseNode, Document, TextNode } from './node';
 export { NodeRelationship, ObjectType } from './types';
-export type { Metadata, RelatedNodeInfo, RelatedNodeType, BaseNodeParams } from './types';
+export type { Metadata, RelatedNodeInfo, RelatedNodeType, BaseNodeParams, TextNodeParams } from './types';
